src/core/api/requirements.txt
boto3==1.24.2
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
# FIXME: Downgraded since core's OIDC server started failing: on token exchange 
# failed: oauth2: cannot fetch token: 400 Bad Request Response: {"error": "invalid_grant"}
#cryptography==41.0.4
cryptography==3.4.8
daphne==4.1.2
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
django-rest-knox==5.0.2
# django-admin-list-filter-dropdown==1.0.3
django-admin-csvexport==2.3
django-autoslug==1.9.8
django-celery-results==2.5.1
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django_extensions==3.2.3
django-filter==24.2
django-grappelli==3.0.8
django-health-check==3.18.3
django-json-widget==1.1.1
django-redis==5.4.0
django-object-actions==3.0.2
django-oauth-toolkit==2.3.0
django-prometheus==2.3.1
django-taggit==6.0.0
docker==6.0.0
dotmap==1.3.30
email-validator==1.3.0
factory-boy==3.2.1
Faker==13.3.4
flower==2.0.0
gql[requests]==3.5.0
GitPython==3.1.43
Jinja2==3.1.2
ipdb==0.13.9
ipython==8.10
kubernetes==29.0.0
prometheus_client==0.20.0
psutil==7.0.0
psycopg2-binary==2.9.9
pydantic==1.10.2
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-django==4.5.2
python-ldap==3.4.3
PyJWT==2.4.0
PyYAML==6.0.2
redis==5.0.7
rich==14.0.0
sentry-sdk
slack-sdk
social-auth-app-django==5.0.0
social-auth-core[openidconnect]==4.1.0
stripe==10.4.0
tenacity==9.0.0
